[fix] Correct default method overrides
Currently default methods are not being handled correctly. If a class
implements a hierarchy of interfaces, then the order of those interfaces
specified by the implementing class determines whether a Miranda or
default is chosen for the vtable slot. Once chosen, the slot is
permanently occupied and cannot be overridden. This is contrary to the
specification that requires the maximally specific super interface
method to be resolved, see 5.4.3.3.3
A test is included, DefaultMethod.